requirements.txt
numpy==2.3.5
pandas==2.3.3
scikit-learn==1.8.0
umap-learn==0.5.9.post2
bertopic==0.17.4
google-genai==1.56.0
matplotlib==3.10.8
japanize-matplotlib==1.1.3
tqdm==4.67.1